Refactor cycle completion notification system
